fix forms and add token

+++ Backend/main.py
@@ -28,7 +28,7 @@ class Response(BaseModel):
 class Create_Task(BaseModel):
     title: str
     description: str
-    status: str
+    status: str | None=None
 
 class update_task(BaseModel):
     title: str
@@ -68,45 +68,72 @@ cursor = connection.cursor()
 def home():
     return {"key": "home page"}
 
+@app.get("/get-user-info")
+def get_user_details(current_user: Annotated[str, Depends(oauth2_scheme)]):
+    try:
+        payload = decode_token(current_user)
+        user_name = payload["user_name"]
+        if not user_name:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"msg": "User not found"})
+
+        fetch_details_query = "SELECT id, username, email FROM users WHERE username=%s"
+        cursor.execute(fetch_details_query, (user_name,))
+        data = cursor.fetchall()
+
+        col_names = [col[0] for col in cursor.description]
+        tasks = [dict(zip(col_names, row)) for row in data]
+        return tasks[0]
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg":f"{e}"})
+
 @app.get("/total-task-length")
-def get_total_length():
-    query = "select count(*) from tasks"
-    cursor.execute(query)
-    length = cursor.fetchone()
-    return length[0]
+def get_total_length(current_user: Annotated[str, Depends(oauth2_scheme)]):
+    try:
+        payload = decode_token(current_user)
+        user_name = payload["user_name"]
+        if not user_name:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"msg": "User not found"})
+        
+        user_id_query = "select id from users where username=%s"
+        cursor.execute(user_id_query, (user_name,))
+        user_id = cursor.fetchone()[0]
+
+        query = "select count(*) from tasks where user_id=%s"
+        cursor.execute(query, (user_id,))
+        length = cursor.fetchone()
+        return length[0]
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg":f"Internal server error {e}"})
 
 @app.get("/get-tasks")
-def get_tasks(page: int | None=None):
+def get_tasks(current_user: Annotated[str, Depends(oauth2_scheme)], page: int | None=None):
+    try:
+        payload = decode_token(current_user)
+        # return payload
+        user_name = payload["user_name"]
+        if not user_name:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"msg": "User not found"})
 
-    query = '''SELECT 
-    tasks.id AS task_id,
-    tasks.title, 
-    tasks.description, 
-    tasks.status, 
-    tasks.created_at, 
-    tasks.updated_at, 
-    users.id AS user_id, 
-    users.username, 
-    users.email
-    FROM tasks
-    JOIN task_assignments ON tasks.id = task_assignments.task_id
-    JOIN users ON task_assignments.user_id = users.id;'''
+        user_id_query = "select id from users where username=%s"
+        cursor.execute(user_id_query, (user_name,))
+        user_id = cursor.fetchone()[0]
 
-   
-    if not page:
-        temp2 = "select * from tasks"
-        cursor.execute(temp2)
-    else:
-        offset = (page - 1) * 6
-        temp_query = f"SELECT * FROM tasks LIMIT 6 OFFSET {offset}"
-        cursor.execute(temp_query)
+        if not page:
+            temp2 = "select * from tasks"
+            cursor.execute(temp2)
+        else:
+            offset = (page - 1) * 6
+            get_tasks_query = "select * from tasks where user_id=%s LIMIT 6 OFFSET %s"
+            cursor.execute(get_tasks_query, (user_id, offset,))
 
-    # cursor.execute(temp_query)
-    col_names = [col[0] for col in cursor.description]
-    data = cursor.fetchall()
+        # cursor.execute(temp_query)
+        col_names = [col[0] for col in cursor.description]
+        data = cursor.fetchall()
 
-    tasks = [dict(zip(col_names, row)) for row in data]
-    return tasks
+        tasks = [dict(zip(col_names, row)) for row in data]
+        return tasks
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg": f"Internal server error {e}"})
 
 @app.post("/register", status_code=status.HTTP_201_CREATED)
 def register(user: User_Register):
@@ -127,7 +154,6 @@ def register(user: User_Register):
 @app.post("/login", status_code=status.HTTP_200_OK)
 def login(user: User_Login):
     try:
-        
         query_password = "SELECT password FROM users WHERE(username=%s)"
         cursor.execute(query_password, (user.user_name,))
         password = cursor.fetchone()
@@ -159,8 +185,8 @@ def create_task(current_user: Annotated[str, Depends(oauth2_scheme)], task: Crea
         cursor.execute(user_id_query, (user_name,))
         user_id = cursor.fetchone()[0]
 
-        query_task = "INSERT INTO Tasks(title, description, status, created_at, user_id) VALUES (%s, %s, %s, %s, %s)"
-        cursor.execute(query_task, (task.title, task.description, task.status, datetime.now(), user_id))
+        query_task = "INSERT INTO Tasks(title, description, created_at, user_id) VALUES (%s, %s, %s, %s)"
+        cursor.execute(query_task, (task.title, task.description, datetime.now(), user_id))
 
         connection.commit()
         return Response(status=True, msg="Task created successfully", status_code=201)
@@ -168,18 +194,32 @@ def create_task(current_user: Annotated[str, Depends(oauth2_scheme)], task: Crea
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg": f"Internal server error {e}"})
 
 @app.put("/task-update", status_code=status.HTTP_200_OK)
-def task_update(task: update_task):
-    update_task_query = "UPDATE tasks SET title = %s, description = %s, status = %s, updated_at =%s WHERE id = %s"
-    cursor.execute(update_task_query, (task.title, task.description, task.status, datetime.now(), task.id))
-    connection.commit()
-    return Response(status=True, msg="Task Updated successfully", status_code=200)
+def task_update(current_user: Annotated[str, Depends(oauth2_scheme)], task: update_task):
+    try:
+        payload = decode_token(current_user)
+        user_name = payload["user_name"]
+        if not user_name:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"msg": "User not found"})
+        update_task_query = "UPDATE tasks SET title = %s, description = %s, status = %s, updated_at =%s WHERE id = %s"
+        cursor.execute(update_task_query, (task.title, task.description, task.status, datetime.now(), task.id))
+        connection.commit()
+        return Response(status=True, msg="Task Updated successfully", status_code=200)
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg": f"Internal server error {e}"})
 
 @app.delete("/delete-task", status_code=status.HTTP_200_OK)
-def delete_task(id: int = Query(...)):
-    delete_query = "DELETE FROM tasks WHERE id=%s"
-    cursor.execute(delete_query, (id, ))
-    connection.commit()
-    return Response(status=True, msg="Task Deleted Successfully", status_code=200)
+def delete_task(current_user: Annotated[str, Depends(oauth2_scheme)],id: int = Query(...)):
+    try:
+        payload = decode_token(current_user)
+        user_name = payload["user_name"]
+        if not user_name:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"msg": "User not found"})
+        delete_query = "DELETE FROM tasks WHERE id=%s"
+        cursor.execute(delete_query, (id, ))
+        connection.commit()
+        return Response(status=True, msg="Task Deleted Successfully", status_code=200)
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail={"msg": f"Internal server error {e}"})
 
 @app.post("/assign-task", status_code=status.HTTP_200_OK)
 def assign_task(current_user: Annotated[str, Depends(oauth2_scheme)], task: task_assignment):
